Extract generated installer definition validation
